LayoutTests/fast/scrolling/fake-mouse-event-throttling.html

<!DOCTYPE html>
<style>
body {
    overflow: scroll;
}
div {
    height: 100px;
}
.pointer {
    cursor: pointer;
    background-color: salmon;
}
.plain {
    background-color: orange;
}
</style>
<script src="../js/resources/js-test-pre.js"></script>
<body onload="runNextTest()">

<script>
for (var i = 0; i < 100; i++) {
    document.write('<div class=pointer></div><div class=plain></div>');
}

if (window.testRunner)
    testRunner.waitUntilDone();
if (window.eventSender)
    eventSender.mouseMoveTo(100, 50);

var numberOfMouseMoves = 0;
document.body.onmousemove = function() {
    var start = Date.now();
    while (Date.now() - start < mouseMoveDuration) {
    }
    numberOfMouseMoves++;

    if (delays.length)
        runNextTest();
    else {
        shouldBe("numberOfMouseMoves", String(expectedNumberOfMouseMoves));
        if (window.testRunner)
            testRunner.notifyDone();
    }

}

var delays = [5, 200];
var expectedNumberOfMouseMoves = delays.length;

function runNextTest()
{
    shouldBe("numberOfMouseMoves", String(expectedNumberOfMouseMoves - delays.length));
    mouseMoveDuration = delays.shift();
    var position = 75 * delays.length;
    window.scrollTo(0, position);
    window.scrollTo(0, position + 25);
    setTimeout(function() {
        window.scrollTo(0, position + 50);
    }, 0);
}
</script>
</body>
