LayoutTests/fast/scrolling/fake-mouse-event-throttling-expected.txt

PASS numberOfMouseMoves is 0
PASS numberOfMouseMoves is 1
PASS numberOfMouseMoves is 2
